backend_django/requirements.txt
# Django and REST Framework
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# WSGI Server
gunicorn==21.2.0

# EEG Processing
mne==1.6.0
numpy==1.26.3
scipy==1.11.4
matplotlib==3.8.2

# Optional: Redis support (uncomment if using Redis)
# django-redis==5.4.0
# redis==5.0.1
